requirements.txt
# ─── Core Framework ────────────────────────────────────────────────────────────
# Python 3.13 compatível — usa apenas pacotes com wheels pré-compilados
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework-simplejwt==5.3.1

# ─── Database ──────────────────────────────────────────────────────────────────
# psycopg2-binary 2.9.10+ tem wheel para Python 3.13 (cp313-win_amd64)
psycopg2-binary>=2.9.10
# UUID PK: UUID suportado nativamente pelo Django desde v1.8

# ─── AWS SDK ───────────────────────────────────────────────────────────────────
# boto3/botocore: puro Python, sem compilação
boto3>=1.34.100
botocore>=1.34.100

# ─── Data Processing ───────────────────────────────────────────────────────────
# pandas 2.2+ tem wheel para Python 3.13
pandas>=2.2.0
openpyxl>=3.1.2
# pyarrow 18+ tem wheel para Python 3.13
pyarrow>=18.0.0
# xlrd para leitura de .xls legados
xlrd>=2.0.1

# ─── Storage & Files ───────────────────────────────────────────────────────────
Pillow>=10.4.0

# ─── API Docs ──────────────────────────────────────────────────────────────────
drf-spectacular>=0.27.2

# ─── Caching & Queue ───────────────────────────────────────────────────────────
celery>=5.4.0
redis>=5.2.0
django-redis>=5.4.0

# ─── Utilities ─────────────────────────────────────────────────────────────────
python-decouple>=3.8
python-dateutil>=2.9.0
requests>=2.32.0
pydantic>=2.9.0
django-extensions>=3.2.3

# ─── Security ──────────────────────────────────────────────────────────────────
# cryptography 43+ tem wheel Python 3.13
cryptography>=43.0.0
python-jose[cryptography]>=3.3.0
PyJWT>=2.9.0

# ─── Logging & Monitoring ──────────────────────────────────────────────────────
structlog>=24.4.0
sentry-sdk>=2.7.0

# ─── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.3.0
pytest-django>=4.9.0
pytest-cov>=5.0.0
factory-boy>=3.3.0
responses>=0.25.0
# moto 5.x tem suporte a Python 3.13
moto[s3,glue,athena,stepfunctions,cognito-idp]>=5.0.20

# ─── Code Quality ──────────────────────────────────────────────────────────────
black>=24.10.0
flake8>=7.1.0
isort>=5.13.2

# ─── Dev Server ────────────────────────────────────────────────────────────────
gunicorn>=22.0.0
whitenoise>=6.7.0
